Fix a bug with the terminal scrolling the entire page
`scrollIntoView` was problematic in the scenario where the terminal was
out of view, as it would scroll the entire page to the terminal.

Adjusting to use `scrollTo`. This PR is slightly bigger as I had to
introduce some forwardRefs to a few text components to get a ref on the
code element itself.